Editor 기능 별 분리
