+++ src/assets/DUMMY_EDITOR_DATA.js
@@ -1,0 +1,25 @@
+// emoji 예시
+const DUMMY_DATA_IMG_SRC =
+    "data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==";
+
+const DUMMY_DATA_TEXT = "default T   ex    t";
+
+const DUMMY_DATA_TITLE = "[dummy] 문서 제목입니다";
+
+/*
+<img 
+    src=${DUMMY_DATA_IMG_SRC} 
+    className=dummy_emoji
+/>
+<div>${DUMMY_DATA_TEXT}</div>
+<div>
+    <div 
+        contentEditable=false
+        className=editor__page_link
+    >
+        ${$EMOJI_EMPTY_DOC}
+        <span className=editor__page_link_title>제목 없음</span>
+    </div>
+</div>
+<div>${DUMMY_DATA_TEXT}</div>
+*/
